requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
greenlet==1.1.2
msgpack==1.0.3
neovim==0.3.1
pynvim==0.4.3
pytz==2021.3
sqlparse==0.4.2
